feat: 剧本会话

+++ src/pages/Main/RightPanel/MaterialItem/index.tsx
@@ -3,13 +3,14 @@ import IconWidget from '@/components/IconWidget'
 import { MoreOutlined, ArrowDownOutlined } from '@ant-design/icons'
 import { ScriptPageList } from '@/api/type'
 import { Flex, Space, Button, Dropdown, message } from 'antd'
-import { useCallback, useContext } from 'react'
+import { useCallback, useContext, useRef } from 'react'
 import classNames from 'classnames'
 import * as api from '@/api/models/main'
 import { downloadFromServer } from '@/utils'
 import type { MenuProps } from 'antd'
 import AntdIcon from '@/components/IconWidget/AntdIcon'
 import { MyContext } from '@/pages/Main'
+import ScriptPreview from '../ScriptPreview'
 interface IMaterialItem {
   data: ScriptPageList // 素材数据
   icon?: string // 素材icon
@@ -17,13 +18,23 @@ interface IMaterialItem {
   handleClick: (val: ScriptPageList) => void
 }
 export default (props: IMaterialItem) => {
-  const { getScriptPageList } = useContext(MyContext)
+  const { getScriptPageList, disabled } = useContext(MyContext)
   const { data } = props
-  const handlePreview = useCallback(() => {
+  const previewRef = useRef<{
+    open: (val: string) => void
+  }>(null)
+  const handlePreview = useCallback(async () => {
     console.log('预览')
+
+    const res = await api.previewScript({
+      scriptId: data.scriptId,
+    })
+    previewRef.current?.open(res)
   }, [])
+
   const handleDel = useCallback(async () => {
     console.log('删除')
+    if (disabled) return
     await api.deleteScript({
       scriptId: data.scriptId,
       projectId: data.projectId,
@@ -56,7 +67,7 @@ export default (props: IMaterialItem) => {
       label: (
         <Flex onClick={handleDel}>
           <AntdIcon style={{ fontSize: '20px' }} icon='delete'></AntdIcon>
-          <span style={{ marginLeft: '8px' }}>删除</span>
+          <span style={{ marginLeft: '8px', cursor: !disabled ? 'pointer' : 'not-allowed' }}>删除</span>
         </Flex>
       ),
     },
@@ -68,29 +79,38 @@ export default (props: IMaterialItem) => {
   }, [])
 
   return (
-    <Flex
-      className={classNames(Styles['material-item'], {
-        [Styles.actived]: data.actived || data.isFinal,
-      })}
-      onClick={() => props.handleClick(props.data)}>
-      <IconWidget className='material-icon' name={props.icon} width={24} height={24} style={{ marginRight: '10px' }} />
-      <Flex className='material-content' align='flex-start' flex={1} vertical={true}>
-        <span className='material-content'>
-          {data?.scriptStyle}-{data.scriptId}
-        </span>
-        <span>{data.modified}</span>
-      </Flex>
-      <div className='material-item-right'>
-        <Space>
-          <Dropdown menu={{ items: downItems }} placement='bottomLeft' arrow={{ pointAtCenter: true }}>
-            <Button type='link' icon={<ArrowDownOutlined />} className={Styles['btn']}></Button>
-          </Dropdown>
+    <>
+      <Flex
+        className={classNames(Styles['material-item'], {
+          [Styles.actived]: data.actived || data.isFinal,
+        })}
+        onClick={() => props.handleClick(props.data)}>
+        <IconWidget
+          className='material-icon'
+          name={props.icon}
+          width={24}
+          height={24}
+          style={{ marginRight: '10px' }}
+        />
+        <Flex className='material-content' align='flex-start' flex={1} vertical={true}>
+          <span className='material-content'>
+            {data?.scriptStyle}-{data.scriptId}
+          </span>
+          <span>{data.modified}</span>
+        </Flex>
+        <div className='material-item-right'>
+          <Space>
+            <Dropdown menu={{ items: downItems }} placement='bottomLeft' arrow={{ pointAtCenter: true }}>
+              <Button type='link' icon={<ArrowDownOutlined />} className={Styles['btn']}></Button>
+            </Dropdown>
 
-          <Dropdown menu={{ items }} placement='bottomLeft' arrow={{ pointAtCenter: true }}>
-            <Button type='link' className={Styles.btn} icon={<MoreOutlined className='material-more' />}></Button>
-          </Dropdown>
-        </Space>
-      </div>
-    </Flex>
+            <Dropdown menu={{ items }} placement='bottomLeft' arrow={{ pointAtCenter: true }}>
+              <Button type='link' className={Styles.btn} icon={<MoreOutlined className='material-more' />}></Button>
+            </Dropdown>
+          </Space>
+        </div>
+      </Flex>
+      <ScriptPreview ref={previewRef} />
+    </>
   )
 }
